Various fixes for test running.
Made sampledata.sh fail if one step fails.
Made run_tests.py use all files related to its dir.
Add check whether data is loaded and server is started.

+++ keystone/test/run_tests.py
@@ -1,26 +1,36 @@
+import os
 import subprocess
 import time
 
 if __name__ == '__main__':
-	#remove pre-existing test databases
-	subprocess.call(['rm', 'keystone.db'])
-	subprocess.call(['rm', 'keystone.token.db'])
+    test_dir = os.path.dirname(__file__)
 
-	# populate the test database
-	subprocess.call(['../../bin/sampledata.sh'])
-	
-	# run the keystone server
-	server = subprocess.Popen(['../../bin/keystone'])
-	
-	# blatent hack.
-	time.sleep(3)
-	
-	# run tests
-	subprocess.call(['python', 'unit/test_keystone.py'])
-	
-	#kill the keystone server
-	server.kill()
-	
-	# remove test databases
-	subprocess.call(['rm', 'keystone.db'])
-	subprocess.call(['rm', 'keystone.token.db'])
+    #remove pre-existing test databases
+    subprocess.call(['rm', os.path.join(test_dir, 'keystone.db')])
+    subprocess.call(['rm', os.path.join(test_dir, 'keystone.token.db')])
+
+    # populate the test database
+    subprocess.check_call([os.path.join(test_dir, '../../bin/sampledata.sh')])
+    
+    try:
+        # run the keystone server
+        server = subprocess.Popen([os.path.join(test_dir,
+                                                '../../bin/keystone')])
+        
+        # blatent hack.
+        time.sleep(3)
+        if server.poll() is not None:
+            print >>sys.stderr, 'Failed to start server'
+            sys.exit(-1)
+        
+        try:
+            # run tests
+            subprocess.check_call(['python',
+                          os.path.join(test_dir, 'unit/test_keystone.py')])
+        finally:
+            #kill the keystone server
+            server.kill()
+    finally:
+        # remove test databases
+        subprocess.call(['rm', os.path.join(test_dir, 'keystone.db')])
+        subprocess.call(['rm', os.path.join(test_dir, 'keystone.token.db')])
